I might have found a really simple answer!
I thought that I would need to be able to detect key up events in order
to make the input a lot less laggy, but simply setting the "pressed" to
true and clearing it when checking if the key is pressed and it is seems
to be "reasonably" effective. I can see this being an issue if you
somehow manage to enter in a lot of keys when the program isn't actually
checking for input, they get buffered, and then in checking for if
certain keys are pressed it first finds you pressed one that indicates you
want to exit (for instance) unintentionally. Might be a problem to solve
down the line but this __seems__ to work for the tank program at least!