Input is responsive + bug fixing (#2)
* I might have found a really simple answer!

I thought that I would need to be able to detect key up events in order
to make the input a lot less laggy, but simply setting the "pressed" to
true and clearing it when checking if the key is pressed and it is seems
to be "reasonably" effective. I can see this being an issue if you
somehow manage to enter in a lot of keys when the program isn't actually
checking for input, they get buffered, and then in checking for if
certain keys are pressed it first finds you pressed one that indicates you
want to exit (for instance) unintentionally. Might be a problem to solve
down the line but this __seems__ to work for the tank program at least!

* Adding a test game for proper testing emulator

Can already see there seems to be some sort of weird issue where it
won't draw the player right in the overworld; will need to look into
that

* Fixed input but weird bugs in cave explorer puzzle

* Set V[0xF] for arithemtic ops AFTER assignment